Add feature-gate for administrators/promotions modals
